requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2026.3
diff-match-patch==20241021
Django==6.0.3
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-environ==0.13.0
django-filter==25.2
django-import-export==4.4.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
kombu==5.6.2
openpyxl==3.1.5
packaging==26.0
pi==0.1.2
pillow==12.1.1
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
reportlab==4.4.10
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
